Return "null" for null keys in tracks, I don't love this
Why? Here's a track:

    { name: "Push It", artistNetwork: ObjectId(…) }

I update it like this:

    { artist: "Sidney’s cover band" }

The server returns this to confirm:

    { name: "Push It", artist: "Sidney’s cover band" }

*Backbone passes that to .set*, and the track model now looks like this:

    { name: "Push It", artist: "Sidney’s cover band", artistNetwork: ObjectId(…) }

Huh, that's not what I wanted. As far as I can tell, there's no way for a
server response to unset keys in a Backbone model, but I don't know whether
to call it a bug in Backbone or not, but I am going to post to the list.

So, for the moment, I think that the best workaround is to pass null for any
key that might have been defined.