Get basic information (tags, length, and md5 sum) from each track.
More Python!

Still need to use the MD5 sum.

You can start editing a track as soon as you start uploading it, we don’t
know its ID3-tagged title or artist until the upload finishes. This introduces
an annoying UX problem: how to make it easy to use the detected title and
artist instead of typing your own. This commit fills in the title and artist
in the editor, but only if that field is empty when the data comes back.